test: パッケージのテストをCIで実行するようにし、腐っていた箇所を修正
- phpunit.xml に Packages スイートを追加し packages/*/tests を対象化
- 旧namespace/旧メソッド名を現行に追随（150件の失敗を解消）
- Laravel依存のミドルウェア/サービステストを api 側へ移設
- 移設前の実装を対象にした化石テストを削除
- PurchaseDto に欠けていた getter を追加
- @test アノテーションを #[Test] 属性へ移行
- ClockUtility の暗黙のfloat→int変換を明示化

+++ packages/nexus-core/src/Utilities/ClockUtility.php
@@ -112,7 +112,7 @@ class ClockUtility
     public static function diffInSeconds(string $dateTimeString): int
     {
         $targetTime = CarbonImmutable::parse($dateTimeString);
-        return self::now()->diffInSeconds($targetTime, false);
+        return (int) self::now()->diffInSeconds($targetTime, false);
     }
 
     /**
@@ -124,7 +124,7 @@ class ClockUtility
     public static function diffInMinutes(string $dateTimeString): int
     {
         $targetTime = CarbonImmutable::parse($dateTimeString);
-        return self::now()->diffInMinutes($targetTime, false);
+        return (int) self::now()->diffInMinutes($targetTime, false);
     }
 
     /**
@@ -136,7 +136,7 @@ class ClockUtility
     public static function diffInHours(string $dateTimeString): int
     {
         $targetTime = CarbonImmutable::parse($dateTimeString);
-        return self::now()->diffInHours($targetTime, false);
+        return (int) self::now()->diffInHours($targetTime, false);
     }
 
     /**
@@ -148,7 +148,7 @@ class ClockUtility
     public static function diffInDays(string $dateTimeString): int
     {
         $targetTime = CarbonImmutable::parse($dateTimeString);
-        return self::now()->diffInDays($targetTime, false);
+        return (int) self::now()->diffInDays($targetTime, false);
     }
 
     /**
